Update remove.sh
Verbesserte deinstallation

+++ wordpress/remove.sh
@@ -17,34 +17,20 @@ function uninstall_single {
     sudo mysql -u root -e "DROP USER IF EXISTS '$install_name'@'localhost';"
 }
 
-# Funktion zum Deinstallieren aller Installationen
-function uninstall_all {
-    # Stopp Apache und MySQL
-    sudo service apache2 stop
-    sudo service mysql stop
+# Funktion zum Fragen, ob Programme entfernt werden sollen
+function ask_remove_programs {
+    read -p "Möchten Sie die folgenden Programme entfernen: apache2 ghostscript libapache2-mod-php mariadb-server php php-bcmath php-curl php-imagick php-intl php-json php-mbstring php-mysql php-xml php-zip (J/N)? " answer
 
-    # Lösche alle WordPress-Installationen und Konfigurationen
-    sudo rm -rf /srv/www/*
-    sudo rm -f /etc/apache2/sites-available/*
-
-    # Datenbanken und Datenbankbenutzer aller Installationen entfernen
-    for db in $(sudo mysql -u root -e "SHOW DATABASES;" | grep -Ev "(Database|information_schema|performance_schema|mysql)"); do
-        sudo mysql -u root -e "DROP DATABASE IF EXISTS $db;"
-        sudo mysql -u root -e "DROP USER IF EXISTS '$db'@'localhost';"
-    done
-
-    # Deinstalliere Apache, MySQL und PHP
-    sudo apt remove --purge -y apache2 mariadb-server php
-    sudo apt autoremove -y
-    sudo apt clean
+    if [[ $answer =~ ^[Jj]$ ]]; then
+        # Deinstalliere die Programme
+        sudo apt remove --purge -y apache2 ghostscript libapache2-mod-php mariadb-server php php-bcmath php-curl php-imagick php-intl php-json php-mbstring php-mysql php-xml php-zip
+        sudo apt autoremove -y
+        sudo apt clean
+    fi
 }
 
-read -p "Möchten Sie eine einzelne Installation (I) oder alle (A) entfernen? " choice
+# Aufruf der Funktion zum Deinstallieren einer einzelnen Installation
+uninstall_single
 
-if [[ $choice =~ ^[Ii]$ ]]; then
-    uninstall_single
-elif [[ $choice =~ ^[Aa]$ ]]; then
-    uninstall_all
-else
-    echo "Ungültige Auswahl. Bitte geben Sie 'I' für einzelne Installation oder 'A' für alle ein."
-fi
+# Aufruf der Funktion zum Fragen, ob Programme entfernt werden sollen
+ask_remove_programs
